fix broken query from last commit, and also make each stats part optional using the new ljmaint arguments feature.
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@82 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ bin/maint/stats.pl
@@ -1,136 +1,161 @@
 #!/usr/bin/perl
 #
 
+use strict;
+use vars qw($dbh %maint);
+
 $maint{'genstats'} = sub
 {
-    &connect_db();
-    my ($sth);
+    my @which = @_;
 
-    print "-I- Getting usage by day in last month...\n";
-    my ($nowtime, $time, $nowdate);
-    $sth = $dbh->prepare("SELECT UNIX_TIMESTAMP(), DATE_FORMAT(NOW(), '%Y-%m-%d')");
-    $sth->execute;
-    ($nowtime, $nowdate) = $sth->fetchrow_array;
-
-    print "Date is: $nowdate\n";
+    unless (@which) { @which = qw(usage users countries 
+				  states gender clients ); }
+    my %do = map { $_, 1, } @which;
     
-    for (my $days_back = 30; $days_back > 0; $days_back--) {
-	print "  going back $days_back days... ";
-	$time = $nowtime - 86400*$days_back;
-	$dbh->do("SET \@d=DATE_FORMAT(FROM_UNIXTIME($time), \"%Y-%m-%d\")");
-	$sth = $dbh->prepare("SELECT COUNT(*) FROM stats WHERE statcat='postsbyday' AND statkey=\@d");
-	$sth->execute;
-	my ($exist) = $sth->fetchrow_array;
-	if ($exist) {
-	    print "exists.\n";
-	} else {
-	    $sth = $dbh->prepare("SELECT \@d, COUNT(*) FROM log WHERE year=YEAR(\@d) AND month=MONTH(\@d) AND day=DAYOFMONTH(\@d)");
-	    $sth->execute;
-	    my ($date, $count) = $sth->fetchrow_array;
-	    print "$date = $count entries\n";
-	    $count += 0;
-	    $dbh->do("REPLACE INTO stats (statcat, statkey, statval) VALUES ('postsbyday', \@d, $count)");	    
-	}
-
-    }
-
-#    print "-I- Getting usage by week...\n";
-
-    print "-I- Getting user stats...\n";
-
+    &connect_db();
+    my $sth;
 
     my %account;
     my %userinfo;
     my %age;
     my %newbyday;
-    my $now = time();
-    my $count;
-    $sth = $dbh->prepare("SELECT COUNT(*) FROM user");
-    $sth->execute;
-    my ($usertotal) = $sth->fetchrow_array;
-    my $pagesize = 1000;
-    my $pages = int($usertotal / $pagesize) + (($usertotal % $pagesize) ? 1 : 0);
-    
-    for (my $page=0; $page < $pages; $page++)
+    my %gender;
+    my %stateus;
+    my %country;
+    my ($nowtime, $time, $nowdate);
+
+    my %to_pop;
+
+    if ($do{'usage'})
     {
-	my $skip = $page*$pagesize;
-	my $first = $skip+1;
-	my $last = $skip+$pagesize;
-	print "  getting records $first-$last...\n";
-	$sth = $dbh->prepare("SELECT DATE_FORMAT(timecreate, '%Y-%m-%d') AS 'datereg', user, paidfeatures, FLOOR((TO_DAYS(NOW())-TO_DAYS(bdate))/365.25) AS 'age', UNIX_TIMESTAMP(timeupdate) AS 'timeupdate', status, allow_getljnews, allow_getpromos FROM user LIMIT $skip,$pagesize");
+	print "-I- Getting usage by day in last month...\n";
+	$sth = $dbh->prepare("SELECT UNIX_TIMESTAMP(), DATE_FORMAT(NOW(), '%Y-%m-%d')");
 	$sth->execute;
-	while (my $rec = $sth->fetchrow_hashref)
-	{
-	    my $co = $rec->{'country'};
-	    if ($co) {
-		$country{$co}++; 
-		if ($co eq "US" && $rec->{'state'}) {
-		    $stateus{$rec->{'state'}}++;
-		}
-	    }
-	    
-	    $account{$rec->{'paidfeatures'}}++;
-
-	    unless ($rec->{'datereg'} eq $nowdate) {
-		$newbyday{$rec->{'datereg'}}++;
-	    }
-
-	    if ($rec->{'age'} > 4 && $rec->{'age'} < 110) {
-		$age{$rec->{'age'}}++;
-	    }
-	    
-	    $userinfo{'total'}++;
-	    $time = $rec->{'timeupdate'};
-	    $userinfo{'updated'}++ if ($time);
-	    $userinfo{'updated_last30'}++ if ($time > $now-60*60*24*30);
-	    $userinfo{'updated_last7'}++ if ($time > $now-60*60*24*7);
-	    $userinfo{'updated_last1'}++ if ($time > $now-60*60*24*1);
-	    
-	    if ($rec->{'status'} eq "A")
-	    {
-		for (qw(allow_getljnews allow_getpromos))
-		{
-		    $userinfo{$_}++ if ($rec->{$_} eq "Y");
-		}
+	($nowtime, $nowdate) = $sth->fetchrow_array;
+	
+	print "Date is: $nowdate\n";
+	
+	for (my $days_back = 30; $days_back > 0; $days_back--) {
+	    print "  going back $days_back days... ";
+	    $time = $nowtime - 86400*$days_back;
+	    $dbh->do("SET \@d=DATE_FORMAT(FROM_UNIXTIME($time), \"%Y-%m-%d\")");
+	    $sth = $dbh->prepare("SELECT COUNT(*) FROM stats WHERE statcat='postsbyday' AND statkey=\@d");
+	    $sth->execute;
+	    my ($exist) = $sth->fetchrow_array;
+	    if ($exist) {
+		print "exists.\n";
+	    } else {
+		$sth = $dbh->prepare("SELECT \@d, COUNT(*) FROM log WHERE year=YEAR(\@d) AND month=MONTH(\@d) AND day=DAYOFMONTH(\@d)");
+		$sth->execute;
+		my ($date, $count) = $sth->fetchrow_array;
+		print "$date = $count entries\n";
+		$count += 0;
+		$dbh->do("REPLACE INTO stats (statcat, statkey, statval) VALUES ('postsbyday', \@d, $count)");	    
 	    }
 	    
 	}
     }
 
-    print "-I- Countries.\n";
-    my %country;
-    $sth = $dbh->prepare("SELECT value, COUNT(*) AS 'count' FROM userprop WHERE upropid=3 AND value<>'' GROUP BY 1 ORDER BY 2");
-    $sth->execute;
-    while ($_ = $sth->fetchrow_hashref) {
-	$country{$_->{'value'}} = $_->{'count'};
+    if ($do{'users'})
+    {
+	$to_pop{'userinfo'} = \%userinfo;
+	$to_pop{'account'} = \%account;
+	$to_pop{'age'} = \%age;
+	$to_pop{'newbyday'} = \%newbyday;
+
+	print "-I- Getting user stats...\n";
+
+	my $now = time();
+	my $count;
+	$sth = $dbh->prepare("SELECT COUNT(*) FROM user");
+	$sth->execute;
+	my ($usertotal) = $sth->fetchrow_array;
+	my $pagesize = 1000;
+	my $pages = int($usertotal / $pagesize) + (($usertotal % $pagesize) ? 1 : 0);
+	
+	for (my $page=0; $page < $pages; $page++)
+	{
+	    my $skip = $page*$pagesize;
+	    my $first = $skip+1;
+	    my $last = $skip+$pagesize;
+	    print "  getting records $first-$last...\n";
+	    $sth = $dbh->prepare("SELECT DATE_FORMAT(timecreate, '%Y-%m-%d') AS 'datereg', user, paidfeatures, FLOOR((TO_DAYS(NOW())-TO_DAYS(bdate))/365.25) AS 'age', UNIX_TIMESTAMP(timeupdate) AS 'timeupdate', status, allow_getljnews, allow_getpromos FROM user LIMIT $skip,$pagesize");
+	    $sth->execute;
+	    while (my $rec = $sth->fetchrow_hashref)
+	    {
+		my $co = $rec->{'country'};
+		if ($co) {
+		    $country{$co}++; 
+		    if ($co eq "US" && $rec->{'state'}) {
+			$stateus{$rec->{'state'}}++;
+		    }
+		}
+		
+		$account{$rec->{'paidfeatures'}}++;
+		
+		unless ($rec->{'datereg'} eq $nowdate) {
+		    $newbyday{$rec->{'datereg'}}++;
+		}
+		
+		if ($rec->{'age'} > 4 && $rec->{'age'} < 110) {
+		    $age{$rec->{'age'}}++;
+		}
+		
+		$userinfo{'total'}++;
+		$time = $rec->{'timeupdate'};
+		$userinfo{'updated'}++ if ($time);
+		$userinfo{'updated_last30'}++ if ($time > $now-60*60*24*30);
+		$userinfo{'updated_last7'}++ if ($time > $now-60*60*24*7);
+		$userinfo{'updated_last1'}++ if ($time > $now-60*60*24*1);
+		
+		if ($rec->{'status'} eq "A")
+		{
+		    for (qw(allow_getljnews allow_getpromos))
+		    {
+			$userinfo{$_}++ if ($rec->{$_} eq "Y");
+		    }
+		}
+		
+	    }
+	}
     }
 
-    print "-I- US States.\n";
-    my %stateus;
-    $sth = $dbh->prepare("SELECT ua.value, COUNT(*) AS 'count' FROM userprop ua, userprop ub WHERE ua.userid=ub.userid AND ua.upropid=4 and ub.upropid=3 and ub.value='US' AND ub.value<>'' GROUP BY 1 ORDER BY 2");
-    $sth->execute;
-    while ($_ = $sth->fetchrow_hashref) {
-	$stateus{$_->{'value'}} = $_->{'count'};
+    if ($do{'countries'})
+    {
+	$to_pop{'county'} = \%country;
+
+	print "-I- Countries.\n";
+	$sth = $dbh->prepare("SELECT value, COUNT(*) AS 'count' FROM userprop WHERE upropid=3 AND value<>'' GROUP BY 1 ORDER BY 2");
+	$sth->execute;
+	while ($_ = $sth->fetchrow_hashref) {
+	    $country{$_->{'value'}} = $_->{'count'};
+	}
     }
 
+    if ($do{'states'}) 
+    {
+	$to_pop{'stateus'} = \%stateus;
 
-    print "-I- Gender.\n";
-    my %gender;
-    $sth = $dbh->prepare("SELECT up.value, COUNT(*) AS 'count' FROM userprop up, userproplist upl WHERE up.upropid=upl.upropid AND upl.name='gender' GROUP BY 1");
-    $sth->execute;
-    while ($_ = $sth->fetchrow_hashref) {
-	$gender{$_->{'value'}} = $_->{'count'};
+	print "-I- US States.\n";
+	$sth = $dbh->prepare("SELECT ua.value, COUNT(*) AS 'count' FROM userprop ua, userprop ub WHERE ua.userid=ub.userid AND ua.upropid=4 and ub.upropid=3 and ub.value='US' AND ub.value<>'' GROUP BY 1 ORDER BY 2");
+	$sth->execute;
+	while ($_ = $sth->fetchrow_hashref) {
+	    $stateus{$_->{'value'}} = $_->{'count'};
+	}
     }
 
-    
-    my %to_pop = ("userinfo" => \%userinfo,
-		  "country" => \%country,
-		  "stateus" => \%stateus,
-		  "age" => \%age,
-		  "gender" => \%gender,
-		  "account" => \%account,
-		  "newbyday" => \%newbyday,
-		  );
+    if ($do{'gender'}) 
+    {
+	$to_pop{'gender'} = \%gender;
+
+	print "-I- Gender.\n";
+	$sth = $dbh->prepare("SELECT up.value, COUNT(*) AS 'count' FROM userprop up, userproplist upl WHERE up.upropid=upl.upropid AND upl.name='gender' GROUP BY 1");
+	$sth->execute;
+	while ($_ = $sth->fetchrow_hashref) {
+	    $gender{$_->{'value'}} = $_->{'count'};
+	}
+    }
+	
     
     foreach my $cat (keys %to_pop)
     {
@@ -148,26 +173,21 @@ $maint{'genstats'} = sub
 
     #### client usage stats
 
-    print "-I- Clients.\n";
-    $sth = $dbh->prepare("SELECT c.client, COUNT(*) FROM clients c, clientusage cu WHERE ".
-			 "c.clientid=cu.clientid AND cu.lastlogin > ".
-			 "DATE_SUB(NOW(), INTERVAL 30 DAY) GROUP BY 1 ORDER BY 2");
-    $sth->execute;
+    if ($do{'clients'})
+    {
+	print "-I- Clients.\n";
+	$sth = $dbh->prepare("SELECT c.client, COUNT(*) AS 'count' FROM clients c, clientusage cu ".
+			     "WHERE c.clientid=cu.clientid AND cu.lastlogin > ".
+			     "DATE_SUB(NOW(), INTERVAL 30 DAY) GROUP BY 1 ORDER BY 2");
+	$sth->execute;
 
-    $dbh->do("DELETE FROM stats WHERE statcat='client'");
-    while ($_ = $sth->fetchrow_hashref) {
-	my $qkey = $dbh->quote($_->{'client'});
-	my $qval = $_->{'count'}+0;
-	$dbh->do("REPLACE INTO stats (statcat, statkey, statval) VALUES ('client', $qkey, $qval)");
+	$dbh->do("DELETE FROM stats WHERE statcat='client'");
+	while ($_ = $sth->fetchrow_hashref) {
+	    my $qkey = $dbh->quote($_->{'client'});
+	    my $qval = $_->{'count'}+0;
+	    $dbh->do("REPLACE INTO stats (statcat, statkey, statval) VALUES ('client', $qkey, $qval)");
+	}
     }
-
-#    print "-I- Overall Active Users\n";
-#    $sth = $dbh->prepare("SELECT u.user, count(*) AS 'count' FROM log l, user u WHERE l.ownerid=u.userid AND u.user NOT LIKE 'test%' GROUP BY user ORDER BY 2 DESC LIMIT 15");
-#    $sth->execute;
-
-#    print "-I- Last Week Active Users\n";
-#    $sth = $dbh->prepare("SELECT u.user, count(*) AS 'count' FROM log  l, user u WHERE l.ownerid=u.userid AND user NOT LIKE 'test%' AND logtime > DATE_SUB(NOW(), INTERVAL 7 DAY) GROUP BY user ORDER BY 2 DESC LIMIT 15");
-#    $sth->execute;
 
     #### dump to text file
     print "-I- Dumping to a text file.\n";
